Implement a max number of servers to round-robin load balancing between.
Before, the client would load-balance between every single server it
knew about. Now, it chooses a subset, and tries to optimize them for
being on different machines.

Algorithm: When we're choosing machines initially or marking machines
as down, walk through the instanceList building a list of indexes of
servers that are good, and a list of indexes of servers that are good
but we've seen their machines before. Then walk through the first
activeSetSize places of our instanceList, and swap a good server into
that place, or if we are out of good servers, swap a less-good
(duplicate machine) server in.