codebeat badge
codebeat is automated code review tool for Swift, Ruby & Go that helps get instant feedback on code quality ...commit-after-commit :)

Is it fine to add codebeat badge to README? 

codebeat suggests "quick wins" in code that could be a nice candidates for a pull request. 

To be fully open and honest. I'm co-founder of codebeat.